requirements.txt
2captcha-python==1.5.1
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
APScheduler==3.9.1.post1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
backtrader==1.9.78.123
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.40.29
botocore==1.40.29
Brotli==1.1.0
browsercookie==0.8.1
buildozer==1.5.0
cachetools==5.0.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.2.0
cloudscraper==1.2.71
cmake==4.0.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
deep-translator==1.11.4
deprecation==2.1.0
dialogflow==2.0.0
discord.py==2.3.2
distlib==0.3.9
distro==1.9.0
docstring_parser==0.17.0
docutils==0.21.2
easyocr==1.7.2
executing==2.1.0
fastapi==0.116.1
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ads==28.4.0
google-api-core==2.28.1
google-auth==2.37.0
google-auth-oauthlib==1.2.3
google-cloud-speech==2.29.0
googleapis-common-protos==1.66.0
gotrue==2.12.3
greenlet==3.2.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
hyperframe==6.1.0
idna==3.11
imageio==2.37.0
instaloader==4.15
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.7.0
jaxlib==0.7.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaitaistruct==0.10
keyboard==0.13.5
keyring==25.6.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
lazy_loader==0.4
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mediapipe==0.10.21
MetaTrader5==5.0.5120
ml_dtypes==0.5.3
more-itertools==10.7.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.0.0
multidict==6.7.0
multitasking==0.0.11
mutagen==1.47.0
narwhals==1.38.2
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
oauthlib==3.3.1
openai==1.93.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pefile==2023.2.7
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.52.0
postgrest==2.24.0
prompt_toolkit==3.0.48
propcache==0.4.1
proto-plus==1.25.0
protobuf==5.29.5
proxyproviders==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydivert==2.1.0
pyee==13.0.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.8.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==22.1
pytimeparse==1.1.8
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.2
realtime==2.24.0
referencing==0.36.2
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.24.0
rsa==4.9
s3transfer==0.14.0
schedule==1.2.2
scikit-image==0.25.2
scipy==1.15.3
screeninfo==0.8.1
selenium==4.32.0
selenium-wire==5.1.0
sentencepiece==0.2.1
setuptools==80.4.0
sh==2.2.2
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.12.0
stack-data==0.6.3
starlette==0.47.2
storage3==2.24.0
streamlit==1.45.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
supafunc==0.10.1
sympy==1.14.0
ta==0.11.0
Telethon==1.40.0
tenacity==9.1.2
tifffile==2025.3.30
TikTokApi==7.2.1
tokenizers==0.22.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
twelvedata==1.2.25
twilio==9.6.0
TwoCaptcha==0.0.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
uvicorn==0.35.0
virtualenv==20.30.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.22.0
yfinance==0.2.55
yt-dlp==2025.11.12
zstandard==0.23.0
